+++ python/rikai/pytorch/vision.py
