+++ python/tests/pytorch/test_torchvision.py
